requirements.txt
pymc>=5.10
arviz>=0.17
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.3
psycopg2-binary>=2.9
